Add files via upload
In this model, i have taken 80% of data for training and 20% data for testing which is the ratio commonly used for training an accurate model. If you have enough data you can follow this ration or else choose the ratio according to your data. It splits the whole data & trains the model with the given ratio of train data(eg. 80%) and test the data with the test ratio specified(20%) which gives the accurate value predicted module.